lifted helper function to `utils`
