+++ projects/gloo/pkg/plugins/upstreamconn/plugin.go
@@ -5,15 +5,15 @@ import (
 
 	envoy_config_cluster_v3 "github.com/envoyproxy/go-control-plane/envoy/config/cluster/v3"
 	envoy_config_core_v3 "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
-	envoy_extensions_http_header_formatters_preserve_case_v3 "github.com/envoyproxy/go-control-plane/envoy/extensions/http/header_formatters/preserve_case/v3"
 	"github.com/golang/protobuf/ptypes/duration"
 	"github.com/golang/protobuf/ptypes/wrappers"
 	"github.com/rotisserie/eris"
 	v1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
 	"github.com/solo-io/gloo/projects/gloo/pkg/api/v1/options/protocol"
 	"github.com/solo-io/gloo/projects/gloo/pkg/plugins"
-	"github.com/solo-io/gloo/projects/gloo/pkg/utils"
 	"github.com/solo-io/solo-kit/pkg/utils/prototime"
+
+	"github.com/solo-io/gloo/projects/gloo/pkg/plugins/utils/httpprotocolhelpers"
 )
 
 var (
@@ -22,8 +22,7 @@ var (
 )
 
 const (
-	ExtensionName      = "upstream_conn"
-	PreserveCasePlugin = "envoy.http.stateful_header_formatters.preserve_case"
+	ExtensionName = "upstream_conn"
 )
 
 type plugin struct{}
@@ -74,7 +73,7 @@ func (p *plugin) ProcessUpstream(params plugins.Params, in *v1.Upstream, out *en
 	}
 
 	if cfg.GetHttp1ProtocolOptions() != nil {
-		http1ProtocolOptions, err := ConvertHttp1ProtocolOptions(*cfg.GetHttp1ProtocolOptions())
+		http1ProtocolOptions, err := httpprotocolhelpers.ConvertHttp1ProtocolOptions(*cfg.GetHttp1ProtocolOptions())
 		if err != nil {
 			return err
 		}
@@ -123,39 +122,6 @@ func convertHttpProtocolOptions(hpo protocol.HttpProtocolOptions) (*envoy_config
 	default:
 		return &envoy_config_core_v3.HttpProtocolOptions{},
 			eris.Errorf("invalid HeadersWithUnderscoresAction %v in CommonHttpProtocolOptions", hpo.GetHeadersWithUnderscoresAction())
-	}
-
-	return out, nil
-}
-
-func ConvertHttp1ProtocolOptions(hpo protocol.Http1ProtocolOptions) (*envoy_config_core_v3.Http1ProtocolOptions, error) {
-	out := &envoy_config_core_v3.Http1ProtocolOptions{}
-
-	if hpo.GetEnableTrailers() {
-		out.EnableTrailers = hpo.GetEnableTrailers()
-	}
-
-	out.OverrideStreamErrorOnInvalidHttpMessage = hpo.GetOverrideStreamErrorOnInvalidHttpMessage()
-
-	if hpo.GetProperCaseHeaderKeyFormat() {
-		out.HeaderKeyFormat = &envoy_config_core_v3.Http1ProtocolOptions_HeaderKeyFormat{
-			HeaderFormat: &envoy_config_core_v3.Http1ProtocolOptions_HeaderKeyFormat_ProperCaseWords_{
-				ProperCaseWords: &envoy_config_core_v3.Http1ProtocolOptions_HeaderKeyFormat_ProperCaseWords{},
-			},
-		}
-	} else if hpo.GetPreserveCaseHeaderKeyFormat() {
-		typedConfig, err := utils.MessageToAny(&envoy_extensions_http_header_formatters_preserve_case_v3.PreserveCaseFormatterConfig{})
-		if err != nil {
-			return nil, err
-		}
-		out.HeaderKeyFormat = &envoy_config_core_v3.Http1ProtocolOptions_HeaderKeyFormat{
-			HeaderFormat: &envoy_config_core_v3.Http1ProtocolOptions_HeaderKeyFormat_StatefulFormatter{
-				StatefulFormatter: &envoy_config_core_v3.TypedExtensionConfig{
-					Name:        PreserveCasePlugin,
-					TypedConfig: typedConfig,
-				},
-			},
-		}
 	}
 
 	return out, nil
